+++ sqldelight-idea-plugin/src/main/kotlin/app/cash/sqldelight/intellij/lang/completion/SqlDelightKeywordCompletionContributor.kt
@@ -87,15 +87,13 @@ class SqlDelightKeywordCompletionContributor : CompletionContributor() {
     ) {
       val originalFile = parameters.originalFile
       val project = originalFile.project
-      val dialect = context.get(dialectKey) ?: run {
+      context.get(dialectKey) ?: run {
         SqlDelightProjectService.getInstance(project).dialect
           .also {
             context.put(dialectKey, it)
           }
       }
-      if (!dialect.isSqlite) {
-        return
-      }
+
       val position = parameters.position
       if (position.node.elementType == SqlTypes.COMMENT) {
         return